requirements.txt
Django>=1.8
MySQL-python
django-ckeditor
sorl-thumbnail